subscription_statistics/requirements.txt
pandas==2.2.2
protobuf==5.27.0
python-dotenv==1.0.1
Requests==2.32.3
